# specs/001-ai/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: 智能標書產生器系統 API
  description: 提供公司資料管理、範本管理、AI內容生成和標書編輯匯出功能的RESTful API
  version: 1.0.0
  contact:
    name: 標書產生器團隊
    email: support@proposal-generator.com
servers:
  - url: http://localhost:3000/api/v1
    description: 開發環境
  - url: https://api.proposal-generator.com/v1
    description: 生產環境
paths:
...
  /companies/basic:
    get:
      tags: [Company]
      summary: 取得公司基本資料
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功取得公司基本資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Company]
      summary: 更新公司基本資料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'

  # 團隊成員 API
...
  /projects:
    get:
      tags: [Projects]
      summary: 取得實績案例列表
      security:
        - bearerAuth: []
      parameters:
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: 標籤篩選
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: 開始日期篩選
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: 結束日期篩選
      responses:
        '200':
          description: 成功取得實績案例列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Projects]
      summary: 新增實績案例
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: 成功新增實績案例
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  # 標書範本 API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: string
        company_name:
          type: string
        tax_id:
          type: string
        capital:
          type: number
        established_date:
          type: string
          format: date
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CompanyUpdate:
      type: object
      properties:
        company_name:
          type: string
          minLength: 2
          maxLength: 255
        tax_id:
          type: string
          pattern: '^[0-9]{8}$'
        capital:
          type: number
          minimum: 0
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
...
    Project:
      type: object
      properties:
        id:
          type: string
        project_name:
          type: string
        client_name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        amount:
          type: number
        scale:
          type: string
        description:
          type: string
        achievements:
          type: string
        tags:
          type: array
          items:
            type: string
        is_public:
          type: boolean
    ProjectCreate:
      type: object
      required: [project_name, description]
      properties:
        project_name:
          type: string
          minLength: 2
          maxLength: 255
        client_name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        amount:
          type: number
          minimum: 0
        scale:
          type: string
        description:
          type: string
        achievements:
          type: string
        tags:
          type: array
          items:
            type: string
        is_public:
          type: boolean
          default: true
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total_pages:
          type: integer
        total_count:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: 請求參數錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...